+++ TOneV2/Code/TOneV2/TOne.WhS.CodePreparation.Entities/SaleCode/MoveCodeInput.cs
@@ -11,7 +11,7 @@ namespace TOne.WhS.CodePreparation.Entities
         public int SellingNumberPlanId { get; set; }
         public int CountryId { get; set; }
 
-        public int ZoneId { get; set; }
+        public long ZoneId { get; set; }
         public string CurrentZoneName { get; set; }
         public string NewZoneName { get; set; }
         public List<string> Codes { get; set; }
